requirements.txt
aiohttp==3.11.11
APScheduler==3.11.0
asyncpraw==7.8.1
firecrawl_py==1.8.0
pydantic==2.10.5
python-dotenv
together==1.3.11
redis
elevenlabs
